Fix iteration order of nodes in CombineMaterialsButton
The order of nodes in a node tree is not well-defined. Because material
hashes were based upon the order of the nodes, this meant that materials
that appeared to be identical could end up with different hashes.

Iteration through nodes to create the material hash now traverses the
node tree via a depth-first search through the node links. Nodes that
are unused by the material will not be included in the hash and the hash
will no longer be affected by the order of the nodes in the node tree.