Speed up and refactor common.fix_zero_length_bones
Whether a bone is considered 'zero-length' is now based on length, which
is how Blender determines whether a bone is considered 'zero-length'.

NumPy and foreach_get/foreach_set are used for performance.
It was already pretty quick due to how few bones there usually are in an
armature, but it's ~6 times faster for a minimal VRChat humanoid (with
all finger bones, so 56 bones), and could be around 18 times faster for
MMD models with rigged clothing and hair (my test case had 440 bones).